10.04

+++ src/com/javarush/task/task10/task1001/Solution.java
@@ -8,9 +8,12 @@ public class Solution {
     public static void main(String[] args) {
         int a = 0;
         int b = (byte) a + 46;
+        //System.out.println(b);
         byte c = (byte) (a * b);
+        //System.out.println(c);
         double f = (char) 1234.15;
-        long d = (short) (a + f / c + b);
+        //System.out.println(f);
+        long d = (long) (a + f / c + b);
         System.out.println(d);
     }
 }
