+++ src/com/javarush/task/task10/task1003/Solution.java
@@ -8,7 +8,7 @@ public class Solution {
     public static void main(String[] args) {
         float f = (float) 128.50;
         int i = (int) f;
-        int b = (int) (i + f);
+        int b = (int) (i + (byte)f);
         System.out.println(b);
     }
 }
