+++ src/com/javarush/task/task10/task1007/Solution.java
@@ -6,10 +6,10 @@ package com.javarush.task.task10.task1007;
 
 public class Solution {
     public static void main(String[] args) {
-        long l = (byte) 1234_564_890L;
-        int x = (byte) 0b1000_1100_1010;
+        long l = (long) 1234_564_890L;
+        int x = (short) 0b1000_1100_1010;
         double m = (byte) 110_987_654_6299.123_34;
-        float f = (byte) l++ + 10 + ++x - (float) m;
+        float f = (long) l++ + 10 + ++x - (float) m;
         l = (long) f / 1000;
         System.out.println(l);
     }
